Refatoração O Que Me Move, Firebase Deploy, Animações e Z-Index
Implementa as refatorações da auditoria AUDIT_06_O_QUE_ME_MOVE.md:
- Estrutura de arquivos: movido para `src/components/sobre/sections/beliefs/`.
- Z-Indexes corrigidos: BeliefBackground z-0, Overlay z-10, Header z-20, Text z-30, Manifesto z-40, GhostCanvas z-50.
- `BeliefOverlay.tsx` integrado no fluxo.
- Refatoração Mobile: saída em "100vw" para fora da tela e âncora vertical `bottom: 20vh`.
- Viewport Reset via `useInView` no componente BeliefsSection.
- Adicionados os workflows automáticos do GitHub Actions para Firebase Hosting.
- Sucesso nos testes e build.
