--- src/components/sobre/3d/index.ts
@@ -1,2 +1,0 @@
-// 3D Components - Ghost Scene
-export { GhostScene } from '@/components/sobre/3d/GhostScene';
