+++ src/components/sobre/sections/AboutBeliefsNoSSR.tsx
@@ -5,8 +5,8 @@ import AboutBeliefsSkeleton from '@/components/sobre/sections/AboutBeliefsSkelet
 
 const AboutBeliefs = dynamic(
   () =>
-    import('@/components/sobre/sections/AboutBeliefsClient').then(
-      (mod) => mod.AboutBeliefsClient
+    import('@/components/sobre/sections/beliefs/BeliefsSection').then(
+      (mod) => mod.BeliefsSection
     ),
   {
     ssr: false,
