+++ src/components/sobre/sections/beliefs/3d/GhostErrorBoundary.tsx
