+++ src/components/sobre/sections/beliefs/3d/index.ts
