+++ src/components/sobre/sections/beliefs/BeliefBackground.tsx
