+++ src/components/sobre/sections/beliefs/BeliefOverlay.tsx
