+++ src/components/sobre/sections/beliefs/BeliefSection.tsx
