+++ src/components/sobre/sections/beliefs/BeliefsScrollContext.tsx
