+++ src/components/sobre/sections/beliefs/CustomCursor.tsx
