+++ src/components/sobre/sections/beliefs/index.ts
